changing return type to double

+++ GetGradient.cu
@@ -4,7 +4,7 @@
 #include "enumerators.h"
 #include <vector>
 
-Eigen::MatrixXf GetGradient(Eigen::MatrixXf Train, Eigen::MatrixXf labels, float sigma, float l)
+Eigen::MatrixXd GetGradient(Eigen::MatrixXf Train, Eigen::MatrixXf labels, float sigma, float l)
 {
     Eigen::MatrixXf gradient(2,1);
     kernel_type type_standard = standard;
@@ -13,7 +13,7 @@ Eigen::MatrixXf GetGradient(Eigen::MatrixXf Train, Eigen::MatrixXf labels, float
     Eigen::MatrixXf Kernel = GetKernel(Train, Train, sigma, l, type_standard);   
     Eigen::MatrixXf dK_dsigma = GetKernel(Train, Train, sigma, l, type_dK_dsigma);   
     Eigen::MatrixXf dK_dl = GetKernel(Train, Train, sigma, l, type_dK_dl);   
-    Eigen::MatrixXf dsigma = 0.5*Multiply(labels.transpose(), Kernel.inverse().transpose());
+    Eigen::MatrixXd dsigma = 0.5*Multiply(labels.transpose(), Kernel.inverse().transpose());
     std::cout << "Kernel" << std::endl;
     for (int row = 0; row != Kernel.inverse().rows(); row++)
     {
